requirements_dev.txt
antspyx>=0.4.2,<0.5.0
innvestigate>=2.1.1,<3.0.0
nilearn>=0.10,<1.0
pytest>=8.3.3,<9.0.0
pytest-cov>=5.0.0,<6.0.0
ruff>=0.6.8,<1.0.0
tensorflow>=2.6,<2.15
tqdm>=4.66.5,<5.0
